replace error ratio analyzer with workspace key metrics analyzer
This Commit also removes the old error rate analyzer.
